// front/src/components/profile/types/ProfileFormData.ts
export interface ProfileFormData {
  firstName: string;
  lastName: string;
  username: string;
  bio: string;
  expertise: string;
  collectif: string;
  avatarUrl: string;
  bannerUrl: string;
  website: string;
  linkedin: string;
  youtube: string;
  github: string;
  spotify: string;
  instagram: string;
  facebook: string;
  accountType: string;
}